refactor(chat): dto <-> entity 변환 mapper 구현
- 기존: dto 내부에서 entity 변환 수행
- 현재: mapper에서 dto <-> entity 변환 전담
