+++ communication-service/src/main/java/com/example/communicationservice/controller/dto/request/ChatMessageSendRequest.java
@@ -1,7 +1,6 @@
 package com.example.communicationservice.controller.dto.request;
 
 import com.example.communicationservice.controller.dto.FileInfo;
-import com.example.communicationservice.entity.ChatMessage;
 import com.example.communicationservice.type.MessageType;
 import jakarta.validation.constraints.NotNull;
 
@@ -21,13 +20,4 @@ public record ChatMessageSendRequest(
     FileInfo file
 
 ) {
-    public ChatMessage toEntity() {
-        return ChatMessage.builder()
-            .roomId(roomId)
-            .senderCode(senderCode)
-            .type(type)
-            .text(text)
-            .file(file != null ? file.toDomain() : null)
-            .build();
-    }
 }
