+++ communication-service/src/main/java/com/example/communicationservice/controller/dto/response/ChatMessageReadResponse.java
@@ -1,7 +1,6 @@
 package com.example.communicationservice.controller.dto.response;
 
 import com.example.communicationservice.controller.dto.FileInfo;
-import com.example.communicationservice.entity.ChatMessage;
 import com.example.communicationservice.type.MessageType;
 
 import java.time.Instant;
@@ -14,14 +13,4 @@ public record ChatMessageReadResponse(
     FileInfo file,
     Instant sentAt
 ) {
-    public static ChatMessageReadResponse from(ChatMessage message) {
-        return new ChatMessageReadResponse(
-            message.getId(),
-            message.getSenderCode(),
-            message.getType(),
-            message.getText(),
-            FileInfo.from(message.getFile()),
-            message.getSentAt()
-        );
-    }
 }
