+++ communication-service/src/main/java/com/example/communicationservice/controller/dto/response/ChatMessageSendResponse.java
@@ -1,7 +1,6 @@
 package com.example.communicationservice.controller.dto.response;
 
 import com.example.communicationservice.controller.dto.FileInfo;
-import com.example.communicationservice.entity.ChatMessage;
 import com.example.communicationservice.type.MessageType;
 
 import java.time.Instant;
@@ -15,15 +14,4 @@ public record ChatMessageSendResponse(
     FileInfo file,
     Instant sentAt
 ) {
-    public static ChatMessageSendResponse from(ChatMessage message) {
-        return new ChatMessageSendResponse(
-            message.getId(),
-            message.getRoomId(),
-            message.getSenderCode(),
-            message.getType(),
-            message.getText(),
-            FileInfo.from(message.getFile()),
-            message.getSentAt()
-        );
-    }
 }
