+++ communication-service/src/main/java/com/example/communicationservice/controller/dto/response/ChatRoomReadResponse.java
@@ -1,7 +1,5 @@
 package com.example.communicationservice.controller.dto.response;
 
-import com.example.communicationservice.entity.ChatRoom;
-
 import java.time.Instant;
 
 public record ChatRoomReadResponse(
@@ -9,11 +7,4 @@ public record ChatRoomReadResponse(
     String name,
     Instant updatedAt
 ) {
-    public static ChatRoomReadResponse from(ChatRoom chatRoom) {
-        return new ChatRoomReadResponse(
-            chatRoom.getId(),
-            chatRoom.getName(),
-            chatRoom.getUpdatedAt()
-        );
-    }
 }
